Fixed incorrect entity id type

+++ src/main/java/ru/bmstu/labs/queriestfidf/repository/DocumentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.bmstu.labs.queriestfidf.model.Document;
 
 @Repository
-public interface DocumentRepository extends JpaRepository<Document, Long> {
+public interface DocumentRepository extends JpaRepository<Document, Integer> {
 }
